requirements.txt
aiogram==3.7.0
alembic==1.13.1
fastapi==0.111.0
loguru==0.7.2
psycopg2-binary==2.9.9
pydantic==2.7.3
python-dotenv==1.0.1
SQLAlchemy==2.0.30
uvicorn==0.30.1
pydantic-settings==2.3.2
pytest==8.2.2
requests==2.32.3